Changes applied in order to conform with code base.
