+++ FastBitmap/FastBitmap.cs
@@ -101,9 +101,9 @@ namespace Hazdryx.Drawing
         /// <param name="image">The image which will be drawn to the bitmap.</param>
         public FastBitmap(Image image) : this(image.Width, image.Height)
         {
-            var bmp = (Bitmap) image;
-            var rect = new Rectangle(0, 0, bmp.Width, bmp.Height);
-            var bmpData = bmp.LockBits(rect, ImageLockMode.ReadWrite, PixelFormat.Format32bppArgb);
+            Bitmap bmp = (Bitmap) image;
+            Rectangle rect = new Rectangle(0, 0, bmp.Width, bmp.Height);
+            BitmapData bmpData = bmp.LockBits(rect, ImageLockMode.ReadWrite, PixelFormat.Format32bppArgb);
 
             Marshal.Copy(bmpData.Scan0, Data, 0, Length);
